backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
greenlet>=3.0.0
alembic>=1.13.1
pydantic>=2.6.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
websockets>=12.0
anthropic>=0.40.0
httpx>=0.26.0
numpy>=1.26.0
reachy-mini>=0.1.0
deepgram-sdk>=3.0.0
openai>=1.0.0
elevenlabs>=2.0.0
google-generativeai>=0.8.0
opencv-python-headless>=4.8.0
boto3>=1.34.0
convex>=0.6.0
paramiko>=3.4.0
scp>=0.14.0
